finish ep24

+++ app/Http/Controllers/ThreadsController.php
@@ -108,6 +108,8 @@ class ThreadsController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         return redirect(route('threads.index'));
